:fire: Renamed debug -> debugoptions to not cause confusion.
