+++ common/src/main/resources/noxesium-common.mixins.json
@@ -6,11 +6,11 @@
   "compatibilityLevel": "JAVA_17",
   "client": [
     "GlowingHotkeyOverrideMixin",
-    "debug.CreativeModeInventoryScreenMixin",
-    "debug.DebugRendererMixin",
-    "debug.EntityRenderDispatcherMixin",
-    "debug.GameRendererPauseMixin",
-    "debug.ScreenTooltipMixin",
+    "debugoptions.CreativeModeInventoryScreenMixin",
+    "debugoptions.DebugRendererMixin",
+    "debugoptions.EntityRenderDispatcherMixin",
+    "debugoptions.GameRendererPauseMixin",
+    "debugoptions.ScreenTooltipMixin",
     "feature.EntityOutlineMixin",
     "feature.FixMovingPistonFlickeringMixin",
     "feature.FixTextDisplayTransparencyMixin",
